Warning修正

+++ bomberStockUI.cpp
@@ -68,7 +68,7 @@ BomberStock::BomberStock() :
 	chargingBG->SetColorObject(SET_COLOR_NOT_COLORED);
 
 	//ƒrƒ…ƒA[
-	viewer = new Viewer3D(SIZE_BOMBER_STOCK_BG.x, SIZE_BOMBER_STOCK_BG.y, D3DXVECTOR2(20.0f, 10.0f));
+	viewer = new Viewer3D((DWORD)SIZE_BOMBER_STOCK_BG.x, (DWORD)SIZE_BOMBER_STOCK_BG.y, D3DXVECTOR2(20.0f, 10.0f));
 	viewer->SetPosition(D3DXVECTOR3((float)SCREEN_WIDTH / 14.5f, -(float)SCREEN_HEIGHT / 100, 0.0f));
 	viewer->SetRotation(0.0f, 30.0f, 0.0f);
 }
