+++ comboUI.cpp
@@ -54,7 +54,7 @@ Combo::Combo()
 	bg->CreateObjectCircle();
 
 	//ƒrƒ…ƒA[
-	viewer = new Viewer3D(SIZE_BACKGROUND_COMBO.x, SIZE_BACKGROUND_COMBO.y, D3DXVECTOR2(20.0f, 10.0f));
+	viewer = new Viewer3D((DWORD)SIZE_BACKGROUND_COMBO.x, (DWORD)SIZE_BACKGROUND_COMBO.y, D3DXVECTOR2(20.0f, 10.0f));
 	viewer->SetPosition(D3DXVECTOR3((float)-SCREEN_WIDTH / 15.0f,  (float)SCREEN_HEIGHT/25.0f, 0.0f));
 	viewer->SetRotation(0.0f, -30.0f, 0.0f);
