+++ scoreUI.cpp
@@ -44,7 +44,7 @@ Score::Score()
 	bg->SetColorObject(SET_COLOR_NOT_COLORED);
 
 	//ƒrƒ…ƒA[
-	viewer = new Viewer3D(SIZE_BG_SCORE.x, SIZE_BG_SCORE.y, D3DXVECTOR2(35.0f, 10.0f));
+	viewer = new Viewer3D((DWORD)SIZE_BG_SCORE.x, (DWORD)SIZE_BG_SCORE.y, D3DXVECTOR2(35.0f, 10.0f));
 	viewer->SetPosition(D3DXVECTOR3((float)-SCREEN_WIDTH/18.0f,(float)SCREEN_HEIGHT / 16.0f, 0.0f));
 	viewer->SetRotation(0.0f, -30.0f, 0.0f);
